[TASK-2-1] divided program into packages
